Polish contribution
