+++ documentation/src/docs/asciidoc/user-guide/extensions.adoc
@@ -489,33 +489,32 @@ constructor invocations, using the `{ExecutableInvoker}` available via the
 
 If multiple implementations of `ParameterResolver` that support the same type are
 registered for a test, a `ParameterResolutionException` will be thrown, with a
-message to indicate that competing resolvers have been discovered. See the following example:
+message to indicate that competing resolvers have been discovered. See the following
+example:
 
 [source,java,indent=0]
-.Conflicting parameter resolution due to duplicate types
+.Conflicting parameter resolution due to multiple resolvers claiming support for integers
 ----
-include::{testDir}/example/extensions/ParameterResolverDuplicateTypeInteger.java[tags=user_guide]
+include::{testDir}/example/extensions/ParameterResolverConflictDemo.java[tags=user_guide]
 ----
 
 If the conflicting `ParameterResolver` implementations are applied to different test
-methods, the methods can be separated into <<writing-tests-nested, `nested tests`>>
-as follows:
+methods as shown in the following example, no conflict occurs.
 
 [source,java,indent=0]
-.Nested parameter resolution to resolve duplicate types
+.Fine-grained registration to avoid conflict
 ----
-include::{testDir}/example/extensions/ParameterResolverNestedClasses.java[tags=user_guide]
+include::{testDir}/example/extensions/ParameterResolverNoConflictDemo.java[tags=user_guide]
 ----
 
-If the conflicting `ParameterResolver` implementations are applied within the same test
-method, implement a custom type or custom annotation as illustrated by
-`{CustomTypeParameterResolver}` and `{CustomAnnotationParameterResolver}`. A custom type
-wraps the duplicate type:
+If the conflicting `ParameterResolver` implementations need to be applied to the same test
+method, you can implement a custom type or custom annotation as illustrated by
+`{CustomTypeParameterResolver}` and `{CustomAnnotationParameterResolver}`, respectively.
 
 [source,java,indent=0]
 .Custom type to resolve duplicate types
 ----
-include::{testDir}/example/extensions/ParameterResolverCustomType.java[tags=user_guide]
+include::{testDir}/example/extensions/ParameterResolverCustomTypeDemo.java[tags=user_guide]
 ----
 
 A custom annotation makes the duplicate type distinguishable from its counterpart:
@@ -523,14 +522,14 @@ A custom annotation makes the duplicate type distinguishable from its counterpar
 [source,java,indent=0]
 .Custom annotation to resolve duplicate types
 ----
-include::{testDir}/example/extensions/ParameterResolverCustomAnnotation.java[tags=user_guide]
+include::{testDir}/example/extensions/ParameterResolverCustomAnnotationDemo.java[tags=user_guide]
 ----
 
-JUnit includes some built-in parameter resolvers that can cause conflicts if an attempt
-is made to redefine their supported types. For example, `{TestInfo}` provides metadata
-about tests. See <<writing-tests-dependency-injection>> for details. Third-party
-frameworks such as Spring may also define parameter resolvers. Apply one of the
-techniques in this section to resolve any conflicts.
+JUnit includes some built-in parameter resolvers that can cause conflicts if a resolver
+attempts to claim their supported types. For example, `{TestInfo}` provides metadata about
+tests. See <<writing-tests-dependency-injection>> for details. Third-party frameworks such
+as Spring may also define parameter resolvers. Apply one of the techniques in this section
+to resolve any conflicts.
 
 Parameterized tests are another potential source of conflict. Ensure that tests annotated
 with `@ParameterizedTest` are not also annotated with `@Test` and see
